+++ mokakbob-api/src/main/java/com/mokakbob/excpeption/DomainExceptionHandler.java
@@ -3,7 +3,7 @@ package com.mokakbob.excpeption;
 
 import com.mokakbob.common.exception.DomainErrorCode;
 import com.mokakbob.common.exception.DomainException;
-import com.mokakbob.excpeption.dto.ErrorResponse;
+import com.mokakbob.excpeption.response.ErrorResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
