+++ mokakbob-api/src/main/java/com/mokakbob/excpeption/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.mokakbob.excpeption.dto;
+package com.mokakbob.excpeption.response;
 
 public record ErrorResponse(
         String customCode,
